Wrap MessagesAvailableView
